% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_functions.R
\name{find_dge_variables}
\alias{find_dge_variables}
\title{Find differentially expressed gene variables of interest}
\usage{
find_dge_variables(object, ident.1, ident.2, my_logfc_threshold,
  my_minpct_threshold)
}
\arguments{
\item{object}{Seurat S4 object. The active identity of the object should
match the ident.1 and ident.2 parameters. You can switch the identity of
your Seurat object with
\code{Idents(object) <- my_ident}.}

\item{ident.1}{Identity class for group-1 for differential expression.}

\item{ident.2}{Optional identity class for group-2 for differential
expression.If differential expression should be compared to all other
cells, leave ident.2 empty.}

\item{my_logfc_threshold}{The log fold-change threshold that will limit
differential expression testing to genes with expression differences
greater than the specified log fold-change threshold. Only genes with a
log fold-change threshold of greater than \code{my_logfc_threshold} will
be returned.}

\item{my_minpct_threshold}{The cutoff for the proportion of cells that
express the gene of interest. Only genes that are expressed in a greater
proportion of cells than the \code{my_minpct_threshold} will be returned.}
}
\description{
\code{find_dge_variables} returns a vector of genes that satisfy differential
expression threshold criteria (log fold-change and min.pct thresholds).
This allows for differential expression to be conducted in a loop and a
progress bar to display within the shiny app indiciating the progress of the
(sometimes lengthy) differential expression. It is modified from the Seurat
FindMarkers() function.
}
\examples{
find_dge_variables(seurat_obj, ident.1 = 2, ident.2 = c(3,4), my_logfc_threshold = 0.25, my_minpct_threshold = 0.5)
find_dge_variables(seurat_obj, ident.1 = 3, my_logfc_threshold = 0.5, my_minpct_threshold = 0.1)
}
