% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_and_export_functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulation of gg colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of colors to return.}
}
\description{
\code{gg_color_hue} Returns a vector containing n unique hex colors
Adapted from stackoverflow John Colby
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
