% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_and_export_functions.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Opposite of in}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{The first vector.}

\item{y}{The second vector.}
}
\description{
Opposite of in
}
\examples{
\dontrun{
c(2,3,4,5) \%not_in\% c(3,4)
}
}
