% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the cellcuratoR shiny app}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
launches the cellcuratoR shiny app
}
\examples{
\dontrun{
launchApp()
}

}
