% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot_functions.R
\name{prepare_violin_data_colors}
\alias{prepare_violin_data_colors}
\title{Reactively generate data (colors and expression profiles) for genes of interest for downstream violin
plotting}
\usage{
prepare_violin_data_colors(my_object, genes_to_investigate,
  dendrogram_input, colors)
}
\arguments{
\item{my_object}{S4 object from which expression data should be extracted.}

\item{genes_to_investigate}{a vector containing character strings of the genes that the user would like
to visualize with violin plots.}

\item{dendrogram_input}{an input dendrogram that relates the expression profiles in each cluster.
See \code{\link{export_shiny_object}} for an example of generating such a dendrogram.}

\item{colors}{a vector of length n_clusters to color the violin plot. The order of colors corresponds to
the order of final_cluster_labels.}
}
\description{
\code{prepare_violin_data_colors} Returns a list of expression data for the requested genes as well as
an ordered vector of colors that will be used to fill the violin plots. This data can then be plotted by the
\code{\link{contruct_violin_plot}} function.
}
\examples{
prepare_violin_data_colors(my_object = seurat_obj, genes_to_investigate = c("RHO", "PDE6A"), dendrogram_input = my_dendrogram, colors = c("red", "green", "orange"))

}
