% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{Shiny app server}
\usage{
shinyAppServer(session, input, output)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
Shiny app server
}
